chore(ZMSKVR-795 GH-1342): refactor AppointmentSelectionView into smaller components
- Split large AppointmentSelectionView into 6 focused components:
  - ProviderView: handles provider selection (multi-provider checkboxes + single provider teaser)
  - CalendarView: handles calendar-based appointment selection with hourly/day-part views
  - ListView: handles accordion-style list view with available days
  - ViewToggle: handles calendar/list view switching with internal state management
  - AppointmentLayout: shared component for appointment time slot display
  - SelectedAppointmentSummary: displays selected appointment details

- Organized components in AppointmentSelection/ folder for better structure
- Moved toggle logic and styles to ViewToggle component for better encapsulation
- Used AppointmentLayout for consistent styling across CalendarView and ListView
- Maintained all existing functionality while improving maintainability
- Fixed recursive update issues with proper event handling and state management
